fixed distance

+++ spotify_lights.py
@@ -44,7 +44,7 @@ def get_beats_info():
 def wave(lights, beat, start_time, min_loudness, max_loudness):
     loudness = (beat["loudness"] - min_loudness) / \
         (max_loudness - min_loudness)
-    distance = loudness * 44
+    distance = int(loudness * 44)
     duration = beat["duration"]
     hsv = (beat["pitch"], 0.99, 0.99)
     for i in range(0, distance):
